% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_url_string}
\alias{.get_url_string}
\title{Format a resource URL as a string}
\usage{
.get_url_string(resource_id)
}
\arguments{
\item{resource_id}{A string for the resource ID from the NHM data portal.}
}
\value{
A string with the url to the data to be downloaded.
}
\description{
Format a resource URL as a string
}
