% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictsr.R
\name{.GetResourceAsData}
\alias{.GetResourceAsData}
\title{From a given resource ID, retrieve the data at the location as a data.frame
or tibble.}
\usage{
.GetResourceAsData(fmt, url_string)
}
\arguments{
\item{fmt}{A string to give the output format, either 'data.frame' or
'tibble'. Defaults to a data frame.}

\item{url_string}{A string giving the URL to download the resource from.}
}
\value{
The site-level summary data in the format specified by 'fmt'.
}
\description{
From a given resource ID, retrieve the data at the location as a data.frame
or tibble.
}
