% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-predicts-data.R
\name{.IsValidPredictsData}
\alias{.IsValidPredictsData}
\title{Check if a PREDICTS extract is valid.}
\usage{
.IsValidPredictsData(df)
}
\arguments{
\item{df}{Dataframe, containing the PREDICTS extract.}
}
\value{
Boolean, \code{TRUE} if the dataframe is valid, \code{FALSE} if not.
}
\description{
A small set of basic checks to ensure that a PREDICTS extract is valid.
These include checking the object is a dataframe, checking all the columns
are valid, and checking that we have a nonzero row count.
}
