% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-predicts-data.R
\name{.ReadPredictsFileCache}
\alias{.ReadPredictsFileCache}
\title{Read the PREDICTS file cache.}
\usage{
.ReadPredictsFileCache(file_predicts, aux_file_predicts, requested_years)
}
\arguments{
\item{file_predicts}{Character, the path to the saved PREDICTS database
extract (as an RDS file).}

\item{aux_file_predicts}{Character, the path to the saved PREDICTS database
auxiliary metadata, saved as a JSON file.}

\item{requested_years}{Numeric vector, the extract years to be saved.}
}
\value{
List, a named list of three elements: 'valid', a boolean indicating if
the cache is valid; 'data', a dataframe containing the cached PREDICTS data;
and 'aux', the auxiliary data loaded from the cache.
}
\description{
This internal helper function returns a list with 3 elements: 'valid', a
boolean indicating if the cache is valid; 'data', a dataframe containing the
cached PREDICTS data; and 'aux', the auxiliary data loaded from the cache.
}
