% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictsr.R
\name{.ReadRDSURL}
\alias{.ReadRDSURL}
\title{Open URL connection and read RDS}
\usage{
.ReadRDSURL(url_string, ...)
}
\arguments{
\item{url_string}{A string (URL) to read the data from.}

\item{...}{Extra arguments passed into \code{readRDS}.}
}
\value{
An R object read in from the RDS file at \code{url_string}.
}
\description{
Open URL connection and read RDS
}
