% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictsr.R
\name{.RequestDataPortal}
\alias{.RequestDataPortal}
\title{Request data from the NHM data portal}
\usage{
.RequestDataPortal(request_body_json, timeout = 600)
}
\arguments{
\item{request_body_json}{A named list giving the body of the request.}

\item{timeout}{Integer giving the time (in seconds) to wait for the request.
Requests in this package usually take < 5s to be processed, so don't set
this to something really high. Default is 600s (10 minutes).}
}
\value{
A named \code{status_json} list giving the status of the request.
Contains \code{status_json$status} which gives if the request was successful or
not. If it was, data can be downloaded from the \code{status_json$urls$direct}
entry.
}
\description{
Request data from the NHM data portal
}
