% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-predicts-data.R
\name{.WritePredictsFileCache}
\alias{.WritePredictsFileCache}
\title{Writes the PREDICTS dataframe to disk.}
\usage{
.WritePredictsFileCache(df, file_predicts, aux_file_predicts, extract)
}
\arguments{
\item{df}{Dataframe to be written to disk}

\item{file_predicts}{Character, path to the desired file - should be an RDS
file.}

\item{aux_file_predicts}{Character, path to where the auxiliary file should
be saved to disk.}

\item{extract}{Numeric vector of release years to be saved.}
}
\value{
TRUE invisibly.
}
\description{
Given a PREDICTS database extract, loaded as an R dataframe, save it to disk
and write the aux JSON file (*.aux.json), which stores metadata for the
object. This includes the release years, the timestamp of when it was saved,
som dimensions, and the SHA-256 hash of the dataframe (computed with
'digest').
}
