% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicts.R
\name{get_predicts_data}
\alias{get_predicts_data}
\title{Read the PREDICTS database into either a data frame or a tibble}
\usage{
get_predicts_data(fmt = "data.frame")
}
\arguments{
\item{fmt}{string, the format to return the data as. Options are 'data.frame'
or 'tibble'.}
}
\value{
A data.frame or tibble containing the full v1.1 PREDICTS database.
}
\description{
This returns the complete PREDICTS database extract from the latest release
(v1.1, 2016). It comprises of 3,278,056 measurements, from 26,194 sampling
locations in 94 countries and representing 47,089 species. The data were
collected as part of the PREDICTS project - Projecting Responses of
Ecological Diversity In Changing Terrestrial Systems.
}
