% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-level.R
\name{get_sitelevel_summaries}
\alias{get_sitelevel_summaries}
\title{Get the site level summaries from the RDS file.}
\usage{
get_sitelevel_summaries(fmt = "data.frame")
}
\description{
Get the site level summaries from the RDS file.
}
