% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictsr.R
\name{GetColumnDescriptions}
\alias{GetColumnDescriptions}
\title{Get a dataframe describing the columns in the PREDICTS database extract.}
\usage{
GetColumnDescriptions(fmt = "data.frame", ...)
}
\arguments{
\item{fmt}{string, the format to return the data as. Options are 'data.frame'
or 'tibble'.}

\item{...}{extra arguments passed to read.csv.}
}
\value{
The column descriptions in the format as specified by 'fmt'.
}
\description{
Get a dataframe describing the columns in the PREDICTS database extract.
}
