% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictsr.R
\name{.GetURLString}
\alias{.GetURLString}
\title{Format a resource URL as a string}
\usage{
.GetURLString(package_id, resource_id)
}
\arguments{
\item{package_id}{A string for the package ID from the NHM data portal.}

\item{resource_id}{A string for the resource ID from the NHM data portal.}
}
\value{
A string with the url to the data to be downloaded.
}
\description{
Format a resource URL as a string
}
