% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictsr.R
\name{.RequestRDSDataFrame}
\alias{.RequestRDSDataFrame}
\title{Request RDS data from the NHM data portal, returning as a dataframe-like.}
\usage{
.RequestRDSDataFrame(status_json, fmt = "data.frame")
}
\arguments{
\item{status_json}{A named list containing the request that was made.}

\item{fmt}{Output format. Either a "data.frame" or "tibble".}
}
\value{
A dataframe or tibble assembled from one or more RDS files returned
by the API.
}
\description{
Request RDS data from the NHM data portal, returning as a dataframe-like.
}
