% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictsr.R
\name{GetColumnDescriptions}
\alias{GetColumnDescriptions}
\title{Get a dataframe describing the columns in the PREDICTS database extract.}
\usage{
GetColumnDescriptions(...)
}
\arguments{
\item{...}{extra arguments passed to read.csv.}
}
\value{
The column descriptions in the format as a dataframe.
}
\description{
Get a dataframe describing the columns in the PREDICTS database extract.
}
\examples{
\donttest{
  descriptions <- GetColumnDescriptions()
}

}
