% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-level.R
\name{GetSitelevelSummaries}
\alias{GetSitelevelSummaries}
\title{Get the site level summaries from the RDS file.}
\usage{
GetSitelevelSummaries(fmt = "data.frame")
}
\arguments{
\item{fmt}{A string to give the output format, either 'data.frame' or
'tibble'. Defaults to a data frame.}
}
\value{
The site-level summary data in the format specified by 'fmt'.
}
\description{
Get the site level summaries from the RDS file.
}
