% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictsr.R
\name{GetSitelevelSummaries}
\alias{GetSitelevelSummaries}
\title{Get the site level summaries from the RDS file.}
\usage{
GetSitelevelSummaries(extract = c(2016, 2022))
}
\arguments{
\item{extract}{Numeric, year/s corresponding to PREDICTS database releases to
download. Options are 2016 or 2022. Defaults to \code{c(2016, 2022)} - the whole
dataset.}
}
\value{
The site-level summary data as a dataframe.
}
\description{
Get the site level summaries from the RDS file.
}
\examples{
\dontrun{
  summaries <- GetSitelevelSummaries()
  summaries_2016 <- GetSitelevelSummaries(extract = 2016)
}

}
