% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictsr-package.R
\docType{package}
\name{predictsr-package}
\alias{predictsr}
\alias{predictsr-package}
\title{predictsr: Access the 'PREDICTS' Biodiversity Database}
\description{
Fetches the 'PREDICTS' database and relevant metadata from the Data Portal at the Natural History Museum, London \url{https://data.nhm.ac.uk}. Data were collated from over 400 existing spatial comparisons of local-scale biodiversity exposed to different intensities and types of anthropogenic pressures, from sites around the world. These data are described in Hudson et al. (2013) \doi{10.1002/ece3.2579}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://biodiversity-futures-lab.github.io/predictsr/}
}

}
\author{
\strong{Maintainer}: Connor Duffin \email{connor.p.duffin@gmail.com}

Other contributors:
\itemize{
  \item The Trustees of The Natural History Museum, London [copyright holder]
}

}
\keyword{internal}
